+++ negotiation/src/main/java/it/eng/negotiation/service/ContractNegotiationConsumerService.java
@@ -175,8 +175,7 @@ public class ContractNegotiationConsumerService extends BaseProtocolService {
      * @param contractNegotiationTerminationMessage
      * @return
      */
-
-    public void handleTerminationResponse(String consumerPid, ContractNegotiationTerminationMessage contractNegotiationTerminationMessage) {
+    public void handleTerminationRequest(String consumerPid, ContractNegotiationTerminationMessage contractNegotiationTerminationMessage) {
     	ContractNegotiation contractNegotiation = findContractNegotiationByPids(contractNegotiationTerminationMessage.getConsumerPid(), contractNegotiationTerminationMessage.getProviderPid());
 
     	stateTransitionCheck(ContractNegotiationState.TERMINATED, contractNegotiation);
@@ -184,24 +183,8 @@ public class ContractNegotiationConsumerService extends BaseProtocolService {
     	ContractNegotiation contractNegotiationTerminated = contractNegotiation.withNewContractNegotiationState(ContractNegotiationState.TERMINATED);
     	contractNegotiationRepository.save(contractNegotiationTerminated);
     	log.info("Contract Negotiation with id {} set to TERMINATED state", contractNegotiation.getId());
-    }
-    
-//    private Offer validateAgreementAgainstOffer(ContractAgreementMessage contractAgreementMessage) {
-//    	return 
-//    	repository.findByProviderPidAndConsumerPid(contractAgreementMessage.getProviderPid(), contractAgreementMessage.getConsumerPid())
-//    		.map(cn -> cn.getOffer())
-//    	.orElseThrow(() -> new OfferNotFoundException(
-//				"Offer with following values from Agreement not found: providerPid " + contractAgreementMessage.getProviderPid() + 
-//				" and consumerPid " + contractAgreementMessage.getConsumerPid() + "and target " + contractAgreementMessage.getAgreement().getTarget()));
-
     	
-//		return offerRepository
-//				.findByConsumerPidAndProviderPidAndTarget(contractAgreementMessage.getConsumerPid(),
-//						contractAgreementMessage.getProviderPid(),
-//						contractAgreementMessage.getAgreement().getTarget())
-//				.orElseThrow(() -> new OfferNotFoundException(
-//						"Offer with following values from Agreement not found: providerPid " + contractAgreementMessage.getProviderPid() + 
-//						" and consumerPid " + contractAgreementMessage.getConsumerPid() + "and target " + contractAgreementMessage.getAgreement().getTarget()));
-//	}
+    	publisher.publishEvent(contractNegotiationTerminationMessage);
+    }
 
 }