+++ negotiation/src/test/java/it/eng/negotiation/model/MockObjectUtil.java
@@ -171,6 +171,16 @@ public class MockObjectUtil {
 			.role(IConstants.ROLE_CONSUMER)
 			.build();
 	
+	public static final ContractNegotiation CONTRACT_NEGOTIATION_REQUESTED_PROIVDER = ContractNegotiation.Builder.newInstance()
+			.consumerPid(MockObjectUtil.CONSUMER_PID)
+			.providerPid(MockObjectUtil.PROVIDER_PID)
+			.callbackAddress(CALLBACK_ADDRESS)
+			.state(ContractNegotiationState.REQUESTED)
+			.offer(OFFER)
+			.assigner(ASSIGNER)
+			.role(IConstants.ROLE_PROVIDER)
+			.build();
+	
 	public static final ContractNegotiation CONTRACT_NEGOTIATION_AGREED = ContractNegotiation.Builder.newInstance()
 			.consumerPid(MockObjectUtil.CONSUMER_PID)
 			.providerPid(MockObjectUtil.PROVIDER_PID)